# swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: MyHomeFinance API
  description: |
    This is small finance program named MyHomeFinance
  contact:
    name: Ivan Pushkin
    url: https://twitter.com/IvanDeveloper
host: localhost:8888
basePath: /
schemes:
  - http
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /note:
    post:
      tags: ["notes"]
      parameters:
        - name: name
          type: string
          in: formData
          required: true
          minLength: 2
          maxLength: 300
          description: length 2-300 chars
        - name: datetime
          type: string
          format: date-time
          in: formData
          description: 'default: current date and time'
        - name: text
          type: string
          in: formData
          description: any text
      responses:
        200:
          description: note created
          schema:
            $ref: '#/definitions/Note'
        400:
          description: invalid parameter(s) supplied
        503:
          description: server is currently unavailable
    get:
      tags: ["notes"]
      parameters:
        - name: id
          in: query
          type: integer
          format: int64
          required: true
          description: note id
          minimum: 1
      responses:
        200:
          description: returns note
          schema:
            $ref: '#/definitions/Note'
        404:
          description: note not found
        503:
          description: server is currently unavailable
    put:
      tags: ["notes"]
      parameters:
        - name: id
          in: query
          type: integer
          format: int64
          required: true
          description: note id
          minimum: 1
        - name: datetime
          in: formData
          type: string
          format: date-time
          description: timestamp
        - name: name
          in: formData
          type: string
          description: note name
          minLength: 2
          maxLength: 300
        - name: text
          in: formData
          type: string
          description: note text
      responses:
        200:
          description: rewrites note
          schema:
            $ref: '#/definitions/Note'
        400:
          description: invalid parameter(s) supplied
        404:
          description: note not found
        503:
          description: server is currently unavailable
    delete:
      tags: ["notes"]
      parameters:
        - name: id
          type: integer
          format: int64
          in: query
          required: true
          description: note id
          minimum: 1
      responses:
        200:
          description: note deleted
        404:
          description: note not found
        503:
          description: server is currently unavailable
  /notes/range:
    get:
      tags: ["notes"]
      parameters:
        - name: date_from
          in: query
          type: string
          required: true
        - name: date_to
          in: query
          type: string
          required: true
      responses:
        200:
          description: list of nodes
          schema:
            type: array
            items:
              $ref: '#/definitions/Note'
        400:
          description: invalid data range supplied
        503:
          description: server is currently unavailable
  /notes/range/match:
    get:
      tags: ["notes"]
      parameters:
        - name: date_from
          in: query
          type: string
          required: true
        - name: date_to
          in: query
          type: string
          required: true
        - name: name
          type: string
          in: formData
          minLength: 1
          maxLength: 300
      responses:
        200:
          description: list of notes
          schema:
            type: array
            items:
              $ref: '#/definitions/Note'
        400:
          description: invalid data range supplied
        503:
          description: server is currently unavailable

  #
  # Inflow
  #
...
definitions:
...
  Note:
    required:
      - id
      - name
      - datetime
    properties:
      id:
        type: integer
        format: int64
        readOnly: true
      name:
        type: string
        minLength: 2
        maxLength: 300
      datetime:
        type: string
        format: date-time
      text:
        type: string

  #
  # Inflow
  #
...